requirements.txt
psycopg2~=2.9.10
ahpy~=2.0
python-dotenv~=1.1.0
pandas~=1.5.3
numpy~=1.26.4